Simplified working with Optionals in VerificationController.
It's now avoided to call Optional.ifPresent()/….get() on optionals but rather use its functional API to handle the result value. Proptypical implementations in VerificationController.verifyTan(…) and ….getTestState(…). I refrained from touching further methods as e.g. generateTan(…) is hard to understand due to its deeply nested ifs and I didn't want to accidentally break things.

The most notable aspect is probably the introduction of the business method VerificationTan.canBeRedeemed(…) as it avoids the controller pulling data out of the object, applying some comparison and deciding based upon that in favor of letting the object tell the client code whether it's eligible to be redeemed.

A couple of other remarks:

- in VerificationController.generateRegistrationToken(…) the call to the service is made even in a code path that could end up in an 400 Bad Request which feels wrong as it seems to execute business logic in an error scenario.
- running the integration tests (VerificationAppTests) against mock services is slightly problematic as the entities work with auto-generated ids and those will stay unpopulated for calls to a repository's save(…) method.