change port settings
This commit changes port settings for the application and the
management interface.

We changed the application port from tomcat default to 8082. This was
done because the default settings of cwa-verification-portal try
to connect to the verification-sever on port 8082 (see
cwa-verification-portal application.yml key
"cwa-verification-server.url"). So we changed it to match that here.

We changed the management interface port from 8081 to 6081. This was
done to facilitate having it run alongside all other backend components
on the same machine for development and testing. This was not possible,
since they all listen on 8081 for the management interface.